requirements.txt
feedparser
urlparse
flask
pyyaml
urllib2
libmc